added force_remove flag to scaledownlist

+++ cloudify_scalelist/workflows.py
@@ -488,6 +488,7 @@ def scaledownlist(ctx, scale_compute=False,
                   scale_node_field_value=u'',
                   all_results=False,
                   node_sequence=None,
+                  force_remove=True,
                   **_):
     if not scale_node_field:
         raise ValueError('You should provide `scale_node_field` for correct'
@@ -533,18 +534,20 @@ def scaledownlist(ctx, scale_compute=False,
     except Exception as e:
         ctx.logger.info('Scale down based on transaction failed: {}'
                         .format(repr(e)))
-        # check list for forced remove
-        removed = []
-        for node in ctx.nodes:
-            for instance in node.instances:
-                if instance.id in instance_ids:
-                    removed.append(instance)
-        _uninstall_instances(ctx=ctx,
-                             graph=ctx.graph_mode(),
-                             removed=removed,
-                             related=[],
-                             ignore_failure=ignore_failure,
-                             node_sequence=node_sequence)
+    
+        if force_remove:
+            # check list for forced remove
+            removed = []
+            for node in ctx.nodes:
+                for instance in node.instances:
+                    if instance.id in instance_ids:
+                        removed.append(instance)
+            _uninstall_instances(ctx=ctx,
+                                graph=ctx.graph_mode(),
+                                removed=removed,
+                                related=[],
+                                ignore_failure=ignore_failure,
+                                node_sequence=node_sequence)
 
         # remove from DB
         if force_db_cleanup:
